+++ imod/mf6/ic.py
@@ -5,12 +5,12 @@ from typing import Any, Optional
 import numpy as np
 
 from imod.common.interfaces.iregridpackage import IRegridPackage
+from imod.common.utilities.regrid import _regrid_package_data
 from imod.logging import init_log_decorator
 from imod.mf6.package import Package
 from imod.mf6.regrid.regrid_schemes import (
     InitialConditionsRegridMethod,
 )
-from imod.mf6.utilities.regrid import RegridderWeightsCache, _regrid_package_data
 from imod.mf6.validation import PKG_DIMS_SCHEMA
 from imod.schemata import (
     AllCoordsValueSchema,
@@ -19,6 +19,7 @@ from imod.schemata import (
     IndexesSchema,
 )
 from imod.typing import GridDataArray
+from imod.util.regrid import RegridderWeightsCache
 
 
 class InitialConditions(Package, IRegridPackage):