+++ imod/mf6/riv.py
@@ -14,9 +14,6 @@ from imod.mf6.drn import Drainage
 from imod.mf6.npf import NodePropertyFlow
 from imod.mf6.regrid.regrid_schemes import RiverRegridMethod
 from imod.mf6.utilities.imod5_converter import regrid_imod5_pkg_data
-from imod.mf6.utilities.regrid import (
-    RegridderWeightsCache,
-)
 from imod.mf6.validation import BOUNDARY_DIMS_SCHEMA, CONC_DIMS_SCHEMA
 from imod.prepare.cleanup import AlignLevelsMode, align_interface_levels, cleanup_riv
 from imod.prepare.topsystem.allocation import ALLOCATION_OPTION, allocate_riv_cells
@@ -46,6 +43,9 @@ from imod.typing.grid import (
     is_planar_grid,
 )
 from imod.util.expand_repetitions import expand_repetitions
+from imod.util.regrid import (
+    RegridderWeightsCache,
+)
 
 
 def set_repeat_stress_if_available(