+++ imod/msw/meteo_grid.py
@@ -9,13 +9,13 @@ import xarray as xr
 
 import imod
 from imod.common.interfaces.iregridpackage import IRegridPackage
+from imod.common.utilities.regrid_method_type import EmptyRegridMethod, RegridMethodType
 from imod.msw.pkgbase import MetaSwapPackage
 from imod.msw.regrid.regrid_schemes import MeteoGridRegridMethod
 from imod.msw.timeutil import to_metaswap_timeformat
 from imod.msw.utilities.common import find_in_file_list
 from imod.msw.utilities.mask import MaskValues
 from imod.typing import Imod5DataDict
-from imod.util.regrid_method_type import EmptyRegridMethod, RegridMethodType
 
 
 class MeteoGrid(MetaSwapPackage, IRegridPackage):