+++ imod/msw/model.py
@@ -10,7 +10,6 @@ import xarray as xr
 
 from imod.mf6.dis import StructuredDiscretization
 from imod.mf6.mf6_wel_adapter import Mf6Wel
-from imod.mf6.utilities.regrid import RegridderWeightsCache
 from imod.msw.copy_files import FileCopier
 from imod.msw.coupler_mapping import CouplerMapping
 from imod.msw.grid_data import GridData
@@ -37,7 +36,7 @@ from imod.msw.utilities.parse import read_para_sim
 from imod.msw.vegetation import AnnualCropFactors
 from imod.typing import Imod5DataDict
 from imod.util.dims import drop_layer_dim_cap_data
-from imod.util.regrid_method_type import RegridderType
+from imod.util.regrid import RegridderType, RegridderWeightsCache
 
 REQUIRED_PACKAGES = (
     GridData,