+++ imod/msw/pkgbase.py
@@ -7,16 +7,16 @@ import numpy as np
 import pandas as pd
 import xarray as xr
 
-from imod.mf6.dis import StructuredDiscretization
-from imod.mf6.mf6_wel_adapter import Mf6Wel
-from imod.mf6.utilities.regrid import (
-    RegridderWeightsCache,
+from imod.common.utilities.regrid import (
     _regrid_like,
 )
+from imod.common.utilities.regrid_method_type import EmptyRegridMethod, RegridMethodType
+from imod.mf6.dis import StructuredDiscretization
+from imod.mf6.mf6_wel_adapter import Mf6Wel
 from imod.msw.fixed_format import format_fixed_width
 from imod.typing import IntArray
 from imod.typing.grid import GridDataArray, GridDataset
-from imod.util.regrid_method_type import EmptyRegridMethod, RegridMethodType
+from imod.util.regrid import RegridderWeightsCache
 
 DataDictType: TypeAlias = dict[str, IntArray | int | str]