+++ imod/tests/test_mf6/test_mf6_rch.py
@@ -10,10 +10,10 @@ import xarray as xr
 
 import imod
 from imod.mf6.dis import StructuredDiscretization
-from imod.mf6.utilities.regrid import RegridderWeightsCache
 from imod.mf6.write_context import WriteContext
 from imod.schemata import ValidationError
 from imod.typing.grid import is_planar_grid, is_transient_data_grid, nan_like
+from imod.util.regrid import RegridderWeightsCache
 
 
 @pytest.fixture(scope="function")